+++ package.json
@@ -18,7 +18,7 @@
     "cookie-parser": "^1.4.5",
     "dayjs": "^1.10.4",
     "dotenv": "^9.0.1",
-    "ejs": "^3.1.5",
+    "ejs": "^3.1.6",
     "express": "^4.15.2",
     "express-session": "^1.17.1",
     "helmet": "^4.6.0",
