+++ views/suggestions.ejs
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>
+    <%= title %>
+  </title>
+  <link rel='stylesheet' href='/stylesheets/style.css' />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+</head>
+
+<body>
+  <h1>
+    <%= title %>
+  </h1>
+  <p>Welcome to <%= title %></p>
+  <ul>
+      <% for (let i = 0; data[i]; ++i) { %>
+        <li>
+            <%= i + 1 %>. <%= data[i] %>
+        </li>
+      <% } %>
+  </ul>
+</body>
